+++ src/calendarjs.js
@@ -1,5 +1,5 @@
 /*
- * Calendar.js Library v1.4.5
+ * Calendar.js Library v1.5.0
  *
  * Copyright 2022 Bunoon
  * Released under the GNU AGPLv3 license
@@ -77,6 +77,7 @@
  * @property    {Object}    onEventsAddedFromJSON                       Specifies an event that will be triggered when events are added from JSON (passes the JSON to the function).
  * @property    {Object}    onDatePickerDateChanged                     Specifies an event that will be triggered when a date is selected in date-picker mode (passes the new date to the function).
  * @property    {Object}    onGroupRemoved                              Specifies an event that will be triggered when a group is removed (passes the group removed to the function).
+ * @property    {Object}    onEventUrlClicked                           Specifies an event that will be triggered when an events Url is clicked (passes the Url to the function).
  */
 
 
@@ -222,6 +223,7 @@
  * @property    {string}    showEmptyDaysInWeekViewText                 The text that should be displayed for the "Show empty days in the week view" label.
  * @property    {string}    showHolidaysInTheDisplaysText               The text that should be displayed for the "Show holidays in the main display and title bars" label.
  * @property    {string}    newEventDefaultTitle                        The default title that should be used for new events (defaults to "* New Event").
+ * @property    {string}    urlErrorMessage                             The error message shown for the "Please enter a valid Url in the 'Url' field (or leave blank)." label.
  */
 
 
@@ -272,6 +274,7 @@
  * @property    {boolean}   useEscapeKeyToExitFullScreenMode            States if the escape key should exit full-screen mode (if enabled, defaults to true).
  * @property    {Object}    minimumDatePickerDate                       States the minimum date that can be selected in DatePicker mode (defaults to null).
  * @property    {Object}    maximumDatePickerDate                       States the minimum date that can be selected in DatePicker mode (defaults to null).
+ * @property    {boolean}   allowHtmlInDisplay                          States if HTML can be used in the display (defaults to false).
  */
 
 
@@ -480,6 +483,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         _element_Tooltip_Repeats = null,
         _element_Tooltip_Description = null,
         _element_Tooltip_Location = null,
+        _element_Tooltip_Url = null,
         _element_Tooltip_ShowTimer = null,
         _element_Tooltip_EventDetails = null,
         _element_DropDownMenu_Day = null,
@@ -790,7 +794,8 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         var headerName = _options.dayHeaderNames[ headerNameIndex ],
             header = createElement( "div", getCellName() );
 
-        header.innerText = headerName;
+        setNodeText( header, headerName );
+
         headerRow.appendChild( header );
 
         header.oncontextmenu = function( e ) {
@@ -1702,9 +1707,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
             plusXEventsText = createElement( "div", "plus-x-events" );
             plusXEventsText.setAttribute( "events", "1" );
-            plusXEventsText.innerText = "+1 " + _options.moreText;
             plusXEventsText.ondblclick = cancelBubble;
             elementDay.appendChild( plusXEventsText );
+
+            setNodeText( plusXEventsText, "+1 " + _options.moreText );
 
             plusXEventsText.onclick = function() {
                 showFullDayView( showFullDayDay, true );
@@ -1713,7 +1719,8 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
             var numberOfEvents = parseInt( plusXEventsText.getAttribute( "events" ) ) + 1;
             plusXEventsText.setAttribute( "events", numberOfEvents.toString() );
-            plusXEventsText.innerText = "+" + numberOfEvents + " " + _options.moreText;
+
+            setNodeText( plusXEventsText, "+" + numberOfEvents + " " + _options.moreText );
         }
     }
 
@@ -1901,7 +1908,8 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         _element_FullDayView_Contents.appendChild( _element_FullDayView_Contents_AllDayEvents );
 
         var allDayText = createElement( "div", "all-day-text" );
-        allDayText.innerText = _options.allDayText;
+        setNodeText( allDayText, _options.allDayText );
+
         _element_FullDayView_Contents_AllDayEvents.appendChild( allDayText );
 
         _element_FullDayView_Contents_Hours = createElement( "div", "contents-events" );
@@ -1958,7 +1966,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
             _element_FullDayView_TodayButton.style.display = "inline-block";
         }
 
-        _element_FullDayView_Title.innerText = "";
+        _element_FullDayView_Title.innerHTML = "";
         _element_FullDayView_DateSelected = new Date( date );
         _element_FullDayView_EventsShown = [];
         _element_FullDayView_Contents_AllDayEvents.style.display = "block";
@@ -2133,33 +2141,32 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         
                 if ( eventDetails.from.getDate() === eventDetails.to.getDate() ) {
                     if ( eventDetails.isAllDay ) {
-                        startTime.innerText = _options.allDayText;
+                        setNodeText( startTime, _options.allDayText );
                     } else {
-                        startTime.innerText = getTimeToTimeDisplay( eventDetails.from, eventDetails.to );
-                        duration.innerText = getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to );
+                        setNodeText( startTime, getTimeToTimeDisplay( eventDetails.from, eventDetails.to ) );
+                        setNodeText( duration, getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to ) );
                     }
                 } else {
 
                     buildDateTimeToDateTimeDisplay( startTime, eventDetails.from, eventDetails.to );
-
-                    duration.innerText = getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to );
+                    setNodeText( duration, getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to ) );
                 }
         
                 if ( isDefinedNumber( eventDetails.repeatEvery ) && eventDetails.repeatEvery > _repeatType.never ) {
                     var repeats = createElement( "div", "repeats" );
-                    repeats.innerText = _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery );
+                    setNodeText( repeats, _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery ) );
                     event.appendChild( repeats );
                 }
         
                 if ( isDefinedStringAndSet( eventDetails.location ) ) {
                     var location = createElement( "div", "location" );
-                    location.innerText = eventDetails.location;
+                    setNodeText( location, eventDetails.location );
                     event.appendChild( location );
                 }
         
                 if ( isDefinedStringAndSet( eventDetails.description ) ) {
                     var description = createElement( "div", "description" );
-                    description.innerText = eventDetails.description;
+                    setNodeText( description, eventDetails.description );
                     event.appendChild( description );
                 }
             }
@@ -2432,7 +2439,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         }
 
         var title = createElement( "div", "title" );
-        title.innerText = _options.allEventsText;
+        setNodeText( title, _options.allEventsText );
         titleBar.appendChild( title );
 
         buildToolbarButton( titleBar, "ib-close", _options.closeTooltipText, function() {
@@ -2533,31 +2540,29 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                     buildDayDisplay( startTime, eventDetails.from, null, " - " + _options.allDayText );
                 } else {
                     buildDayDisplay( startTime, eventDetails.from, null, " - " + getTimeToTimeDisplay( eventDetails.from, eventDetails.to ) );
-
-                    duration.innerText = getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to );
+                    setNodeText( duration, getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to ) );
                 }
             } else {
 
                 buildDateTimeToDateTimeDisplay( startTime, eventDetails.from, eventDetails.to );
-
-                duration.innerText = getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to );
+                setNodeText( duration, getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to ) );
             }
     
             if ( isDefinedNumber( eventDetails.repeatEvery ) && eventDetails.repeatEvery > _repeatType.never ) {
                 var repeats = createElement( "div", "repeats" );
-                repeats.innerText = _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery );
+                setNodeText( repeats, _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery ) );
                 event.appendChild( repeats );
             }
     
             if ( isDefinedStringAndSet( eventDetails.location ) ) {
                 var location = createElement( "div", "location" );
-                location.innerText = eventDetails.location;
+                setNodeText( location, eventDetails.location );
                 event.appendChild( location );
             }
     
             if ( isDefinedStringAndSet( eventDetails.description ) ) {
                 var description = createElement( "div", "description" );
-                description.innerText = eventDetails.description;
+                setNodeText( description, eventDetails.description );
                 event.appendChild( description );
             }
     
@@ -2589,7 +2594,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
             makeAreaDroppable( month, date.getFullYear(), date.getMonth() );
 
             var header = createElement( "div", "header" );
-            header.innerText = _options.monthNames[ date.getMonth() ] + " " + date.getFullYear();
+            setNodeText( header, _options.monthNames[ date.getMonth() ] + " " + date.getFullYear() );
             header.ondblclick = expandFunction;
             month.appendChild( header );
 
@@ -2889,33 +2894,32 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
     
             if ( eventDetails.from.getDate() === eventDetails.to.getDate() ) {
                 if ( eventDetails.isAllDay ) {
-                    startTime.innerText = _options.allDayText;
+                    setNodeText( startTime, _options.allDayText );
                 } else {
-                    startTime.innerText = getTimeToTimeDisplay( eventDetails.from, eventDetails.to );
-                    duration.innerText = getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to );
+                    setNodeText( startTime, getTimeToTimeDisplay( eventDetails.from, eventDetails.to ) );
+                    setNodeText( duration, getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to ) );
                 }
             } else {
 
                 buildDateTimeToDateTimeDisplay( startTime, eventDetails.from, eventDetails.to );
-
-                duration.innerText = getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to );
+                setNodeText( duration, getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to ) );
             }
     
             if ( isDefinedNumber( eventDetails.repeatEvery ) && eventDetails.repeatEvery > _repeatType.never ) {
                 var repeats = createElement( "div", "repeats" );
-                repeats.innerText = _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery );
+                setNodeText( repeats, _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery ) );
                 event.appendChild( repeats );
             }
     
             if ( isDefinedStringAndSet( eventDetails.location ) ) {
                 var location = createElement( "div", "location" );
-                location.innerText = eventDetails.location;
+                setNodeText( location, eventDetails.location );
                 event.appendChild( location );
             }
     
             if ( isDefinedStringAndSet( eventDetails.description ) ) {
                 var description = createElement( "div", "description" );
-                description.innerText = eventDetails.description;
+                setNodeText( description, eventDetails.description );
                 event.appendChild( description );
             }
     
@@ -3591,7 +3595,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         }
 
         _element_DropDownMenu_Event_OpenUrl = buildMenuItemWithIcon( _element_DropDownMenu_Event, "ib-arrow-top-right-icon", _options.openUrlText, function() {
-            _window.open( _element_DropDownMenu_Event_EventDetails.url, _options.urlWindowTarget );
+            openEventUrl( _element_DropDownMenu_Event_EventDetails.url );
         } );
     }
 
@@ -3671,7 +3675,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         menuItem.appendChild( createElement( "div", iconCSS ) );
 
         var menuText = createElement( "div", "menu-text" );
-        menuText.innerText = text;
+        setNodeText( menuText, text );
         menuItem.appendChild( menuText );
 
         if ( isBold ) {
@@ -4035,9 +4039,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         selectTab( _element_EventEditorDialog );
 
         if ( isDefined( eventDetails ) ) {
+            setNodeText( _element_EventEditorDialog_TitleBar, _options.editEventTitle );
+
             _element_EventEditorDialog_OKButton.value = _options.updateText;
             _element_EventEditorDialog_RemoveButton.style.display = "block";
-            _element_EventEditorDialog_TitleBar.innerText = _options.editEventTitle;
             _element_EventEditorDialog_EventDetails = eventDetails;
             _element_EventEditorDialog_TimeFrom.value = toFormattedTime( eventDetails.from );
             _element_EventEditorDialog_TimeTo.value = toFormattedTime( eventDetails.to );
@@ -4104,9 +4109,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 today.setMinutes( date.getMinutes() );
             }
 
+            setNodeText( _element_EventEditorDialog_TitleBar, _options.addEventTitle );
+
             _element_EventEditorDialog_OKButton.value = _options.addText;
             _element_EventEditorDialog_RemoveButton.style.display = "none";
-            _element_EventEditorDialog_TitleBar.innerText = _options.addEventTitle;
             _element_EventEditorDialog_EventDetails = {};
             _element_EventEditorDialog_IsAllDay.checked = false;
             _element_EventEditorDialog_ShowAlerts.checked = true;
@@ -4155,9 +4161,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
     }
 
     function setEventEditingDialogInDuplicateMode() {
+        setNodeText( _element_EventEditorDialog_TitleBar, _options.addEventTitle );
+
         _element_EventEditorDialog_OKButton.value = _options.addText;
         _element_EventEditorDialog_RemoveButton.style.display = "none";
-        _element_EventEditorDialog_TitleBar.innerText = _options.addEventTitle;
         _element_EventEditorDialog_EventDetails = cloneEventDetails( _element_EventEditorDialog_EventDetails );
 
         buildToolbarButton( _element_EventEditorDialog_TitleBar, "ib-close", _options.closeTooltipText, eventDialogEvent_Cancel, true );
@@ -4166,7 +4173,8 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
     function eventDialogEvent_OK() {
         var fromTime = _element_EventEditorDialog_TimeFrom.value.split( ":" ),
             toTime = _element_EventEditorDialog_TimeTo.value.split( ":" ),
-            title = trimString( _element_EventEditorDialog_Title.value );
+            title = trimString( _element_EventEditorDialog_Title.value ),
+            url = trimString( _element_EventEditorDialog_Url.value );
 
         if ( fromTime.length < 2 ) {
             showEventDialogErrorMessage( _options.fromTimeErrorMessage, _element_EventEditorDialog_TimeFrom );
@@ -4174,6 +4182,8 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
             showEventDialogErrorMessage( _options.toTimeErrorMessage, _element_EventEditorDialog_TimeTo );
         } else if ( title === "" ) {
             showEventDialogErrorMessage( _options.titleErrorMessage, _element_EventEditorDialog_Title );
+        } else if ( url.length > 0 && !isValidUrl( url ) ) {
+            showEventDialogErrorMessage( _options.urlErrorMessage, _element_EventEditorDialog_Title );
         }
         else {
 
@@ -4183,7 +4193,6 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 location = trimString( _element_EventEditorDialog_Location.value ),
                 group = trimString( _element_EventEditorDialog_Group.value ),
                 repeatEnds = getSelectedDate( _element_EventEditorRepeatOptionsDialog_RepeatEnds, null ),
-                url = trimString( _element_EventEditorDialog_Url.value ),
                 repeatEveryCustomValue = parseInt( _element_EventEditorDialog_RepeatEvery_Custom_Value.value );
 
             setTimeOnDate( fromDate, _element_EventEditorDialog_TimeFrom.value );
@@ -4293,7 +4302,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
     }
 
     function showEventDialogErrorMessage( message, input ) {
-        _element_EventEditorDialog_ErrorMessage.innerText = message;
+        setNodeText( _element_EventEditorDialog_ErrorMessage, message );
         _element_EventEditorDialog_ErrorMessage.style.display = "block";
 
         input.focus();
@@ -4358,7 +4367,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         _document.body.appendChild( _element_EventEditorColorsDialog );
 
         var titleBar = createElement( "div", "title-bar" );
-        titleBar.innerText = _options.selectColorsText;
+        setNodeText( titleBar, _options.selectColorsText );
         _element_EventEditorColorsDialog.appendChild( titleBar );
 
         buildToolbarButton( titleBar, "ib-close", _options.closeTooltipText, eventColorsDialogEvent_Cancel, true );
@@ -4431,7 +4440,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         _document.body.appendChild( _element_EventEditorRepeatOptionsDialog );
 
         var titleBar = createElement( "div", "title-bar" );
-        titleBar.innerText = _options.repeatOptionsTitle;
+        setNodeText( titleBar, _options.repeatOptionsTitle );
         _element_EventEditorRepeatOptionsDialog.appendChild( titleBar );
 
         buildToolbarButton( titleBar, "ib-close", _options.closeTooltipText, eventRepeatOptionsDialogEvent_Cancel, true );
@@ -4557,8 +4566,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
         _openDialogs.push( false );
         _element_ConfirmationDialog.style.display = "block";
-        _element_ConfirmationDialog_TitleBar.innerText = title;
-        _element_ConfirmationDialog_Message.innerText = message;
+
+        setNodeText( _element_ConfirmationDialog_TitleBar, title );
+        setNodeText( _element_ConfirmationDialog_Message, message );
+
         _element_ConfirmationDialog_YesButton.onclick = hideConfirmationDialog;
         _element_ConfirmationDialog_NoButton.onclick = hideConfirmationDialog;
         _element_ConfirmationDialog_YesButton.addEventListener( "click", onYesEvent );
@@ -4597,7 +4608,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         _document.body.appendChild( _element_SelectExportTypeDialog );
 
         var titleBar = createElement( "div", "title-bar" );
-        titleBar.innerText = _options.selectExportTypeTitle;
+        setNodeText( titleBar, _options.selectExportTypeTitle );
         _element_SelectExportTypeDialog.appendChild( titleBar );
 
         buildToolbarButton( titleBar, "ib-close", _options.closeTooltipText, hideSelectExportTypeDialog, true );
@@ -4685,7 +4696,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         _document.body.appendChild( _element_SearchDialog );
 
         var titleBar = createElement( "div", "title-bar" );
-        titleBar.innerText = _options.searchEventsTitle;
+        setNodeText( titleBar, _options.searchEventsTitle );
         titleBar.onmousedown = searchOnTitleBarMouseDown;
         titleBar.onmouseup = searchOnTitleBarMouseUp;
         titleBar.oncontextmenu = searchOnTitleBarMouseUp;
@@ -5080,7 +5091,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         _document.body.appendChild( _element_ConfigurationDialog );
 
         var titleBar = createElement( "div", "title-bar" );
-        titleBar.innerText = _options.configurationTitleText;
+        setNodeText( titleBar, _options.configurationTitleText );
         _element_ConfigurationDialog.appendChild( titleBar );
 
         buildToolbarButton( titleBar, "ib-close", _options.closeTooltipText, configurationDialogEvent_Cancel, true );
@@ -5335,6 +5346,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
             _element_Tooltip_Repeats = createElement( "div", "repeats" );
             _element_Tooltip_Description = createElement( "div", "description" );
             _element_Tooltip_Location = createElement( "div", "location" );
+            _element_Tooltip_Url = createElement( "div", "url" );
 
             _element_Tooltip_TitleButtons_CloseButton.onclick = hideTooltip;
             _element_Tooltip_TitleButtons_EditButton.onclick = function() {
@@ -5360,7 +5372,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                     _element_Tooltip.className = text === "" ? "calendar-tooltip-event" : "calendar-tooltip";
 
                     if ( text !== "" ) {
-                        _element_Tooltip.innerText = text;
+                        setNodeText( _element_Tooltip, text );
                     } else {
 
                         _element_Tooltip.onmousemove = cancelBubble;
@@ -5373,6 +5385,21 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                         _element_Tooltip.appendChild( _element_Tooltip_Date );
                         _element_Tooltip.appendChild( _element_Tooltip_TotalTime );
 
+                        if ( isDefinedStringAndSet( eventDetails.url ) ) {
+                            setNodeText( _element_Tooltip_Url, getShortUrlString( eventDetails.url ) );
+                            addNode( _element_Tooltip, _element_Tooltip_Url );
+
+                            _element_Tooltip_Url.onclick = function( e ) {
+                                cancelBubble( e );
+                                openEventUrl( eventDetails.url );
+                            };
+
+                        } else {
+                            _element_Tooltip_Url.innerHTML = "";
+                            _element_Tooltip_Url.onclick = null;
+                            removeNode( _element_Tooltip, _element_Tooltip_Url );
+                        }
+
                         var repeatEvery = getNumber( eventDetails.repeatEvery );
                         if ( repeatEvery > _repeatType.never ) {
                             var icon = createElement( "div", "ib-refresh-medium ib-no-hover ib-no-active" );
@@ -5383,41 +5410,40 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                         _element_Tooltip_Title.innerHTML += stripHTMLTagsFromText( eventDetails.title );
 
                         if ( isDefinedNumber( eventDetails.repeatEvery ) && eventDetails.repeatEvery > _repeatType.never ) {
-                            _element_Tooltip_Repeats.innerText = _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery );
+                            setNodeText( _element_Tooltip_Repeats, _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery ) );
                             addNode( _element_Tooltip, _element_Tooltip_Repeats );
                         } else {
-                            _element_Tooltip_Repeats.innerText = "";
+                            _element_Tooltip_Repeats.innerHTML = "";
                             removeNode( _element_Tooltip, _element_Tooltip_Repeats );
                         }
 
                         if ( isDefinedStringAndSet( eventDetails.location ) ) {
-                            _element_Tooltip_Location.innerText = eventDetails.location;
+                            setNodeText( _element_Tooltip_Location, eventDetails.location );
                             addNode( _element_Tooltip, _element_Tooltip_Location );
                         } else {
-                            _element_Tooltip_Location.innerText = "";
+                            _element_Tooltip_Location.innerHTML = "";
                             removeNode( _element_Tooltip, _element_Tooltip_Location );
                         }
     
                         if ( isDefinedStringAndSet( eventDetails.description ) ) {
-                            _element_Tooltip_Description.innerText = eventDetails.description;
+                            setNodeText( _element_Tooltip_Description, eventDetails.description );
                             addNode( _element_Tooltip, _element_Tooltip_Description );
                         } else {
-                            _element_Tooltip_Description.innerText = "";
+                            _element_Tooltip_Description.innerHTML = "";
                             removeNode( _element_Tooltip, _element_Tooltip_Description );
                         }
         
                         if ( eventDetails.from.getDate() === eventDetails.to.getDate() ) {
                             if ( eventDetails.isAllDay ) {
-                                _element_Tooltip_Date.innerText = _options.allDayText;
+                                setNodeText( _element_Tooltip_Date, _options.allDayText );
                             } else {
-                                _element_Tooltip_Date.innerText = getTimeToTimeDisplay( eventDetails.from, eventDetails.to );
-                                _element_Tooltip_TotalTime.innerText = getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to );
+                                setNodeText( _element_Tooltip_Date, getTimeToTimeDisplay( eventDetails.from, eventDetails.to ) );
+                                setNodeText( _element_Tooltip_TotalTime, getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to ) );
                             }
                         } else {
 
                             buildDateTimeToDateTimeDisplay( _element_Tooltip_Date, eventDetails.from, eventDetails.to );
-
-                            _element_Tooltip_TotalTime.innerText = getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to );
+                            setNodeText( _element_Tooltip_TotalTime, getFriendlyTimeBetweenTwoDate( eventDetails.from, eventDetails.to ) );
                         }
                     }
 
@@ -5477,7 +5503,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         var className = "tab tab-control" + ( selected ? "-selected" : "" ),
             tab = createElement( "div", className );
 
-        tab.innerText = text;
+        setNodeText( tab, text );
         container.appendChild( tab );
 
         tab.onclick = function () {
@@ -5580,7 +5606,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
             if ( url === "" ) {
                 showEventEditingDialog( event );
             } else {
-                _window.open( url, _options.urlWindowTarget );
+                openEventUrl( url );
             }
 
             triggerOptionsEventWithData( "onNotificationClicked", event );
@@ -5607,6 +5633,12 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 action();
             }
         }
+    }
+
+    function openEventUrl( url ) {
+        _window.open( url, _options.urlWindowTarget );
+
+        triggerOptionsEvent( "onEventUrlClicked", url );
     }
 
 
@@ -5704,7 +5736,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
     function createTextHeaderElement( container, text ) {
         var element = createElement( "p" );
-        element.innerText = text;
+        setNodeText( element, text );
         container.appendChild( element );
     }
 
@@ -5713,8 +5745,8 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
         var element = createElement( "span", className ),
             isEventDefined = isDefined( event );
-
-        element.innerText = text;        
+ 
+        setNodeText( element, text );    
         container.appendChild( element );
 
         if ( isEventDefined ) {
@@ -5900,10 +5932,16 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
     }
 
     function stripHTMLTagsFromText( text ) {
-        var div = createElement( "div" );
-        div.innerHTML = text;
+        var result = text;
 
-        return div.innerText;
+        if ( !_options.allowHtmlInDisplay ) {
+            var div = createElement( "div" );
+            div.innerHTML = text;
+    
+            result = div.innerText;
+        }
+
+        return result;
     }
 
     function getStyleValueByName( element, stylePropertyName ) {
@@ -5917,6 +5955,14 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         }                     
 
         return value;
+    }
+
+    function setNodeText( element, text ) {
+        if ( !_options.allowHtmlInDisplay ) {
+            element.innerText = stripHTMLTagsFromText( text );
+        } else {
+            element.innerHTML = text;
+        }
     }
 
 
@@ -5955,7 +6001,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         container.appendChild( lineContents );
 
         var label = createElement( "label", "checkbox" + extraClassName );
-        label.innerText = labelText;
+        setNodeText( label, labelText );
         lineContents.appendChild( label );
 
         var input = createElement( "input", null, "checkbox" );
@@ -6046,6 +6092,20 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         return data.substring( data.length - end.length, data.length ) === end;
     }
 
+    function getShortUrlString( url, maxLength ) {
+        var result = url;
+
+        maxLength = isDefined( maxLength ) ? maxLength : 30;
+
+        if ( url.length > maxLength ) {
+            var sideLength = maxLength % 2 === 0 ? maxLength / 2 : ( maxLength - 1 ) / 2;
+
+            result = url.substring( 0, sideLength ) + "..." + url.substring( url.length - sideLength );
+        }
+
+        return result;
+    }
+
 
     /*
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -6095,6 +6155,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
     function isDefinedDate( object ) {
         return isDefinedObject( object ) && object instanceof Date;
+    }
+
+    function isValidUrl( url ) {
+        return /^(?:(?:(?:https?|ftp):)?\/\/)(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))(?::\d{2,5})?(?:[/?#]\S*)?$/i.test( url );
     }
 
 
@@ -7183,7 +7247,8 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 title = getString( event.title ),
                 description = getString( event.description ),
                 location = getString( event.location ),
-                group = getString( event.group );
+                group = getString( event.group ),
+                url = getString( event.url );
 
             if ( !_events.hasOwnProperty( storageDate ) ) {
                 _events[ storageDate ] = {};
@@ -7213,6 +7278,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
                 if ( _options.maximumEventGroupLength > 0 && group !== "" && group.length > _options.maximumEventGroupLength ) {
                     event.group = event.group.substring( 0, _options.maximumEventGroupLength );
+                }
+
+                if ( url !== "" && !isValidUrl( url ) ) {
+                    event.url = "";
                 }
 
                 if ( !isDefined( event.created ) ) {
@@ -7582,7 +7651,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "1.4.5";
+        return "1.5.0";
     };
 
 
@@ -7887,6 +7956,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
         if ( !isDefinedDate( _options.maximumDatePickerDate ) ) {
             _options.maximumDatePickerDate = null;
+        }
+
+        if ( !isDefinedBoolean( _options.allowHtmlInDisplay ) ) {
+            _options.allowHtmlInDisplay = false;
         }
 
         setTranslationStringOptions();
@@ -8529,6 +8602,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
         if ( !isDefinedString( _options.newEventDefaultTitle ) ) {
             _options.newEventDefaultTitle = "* New Event";
+        }
+
+        if ( !isDefinedString( _options.urlErrorMessage ) ) {
+            _options.urlErrorMessage = "Please enter a valid Url in the 'Url' field (or leave blank).";
         }
     }